project: Update candle.js to v0.5.0 and remove jsdom dependency
candle.js v0.5.0 implements XML parsing and serialize routine.
This enables to remove jsdom dependency to handle XML files and
results in parsing and serializing XML file x50 faster.

In addition, if node-canvas is installed globally,
this script can detect and use it to output png image file if necessary.