fix crossing loop bug by rebuilding graph
This is a very heavy-handed fix, which might be noticeably slow.
It might also be wrong to repaint in the postupdate border hook;
I had to explicitly check for presence of a drawing context to
not break other tests.

Still, it should fix the bug for now; should be followed up by
a more thorough fix that either computes the connectivity
correctly incrementally, or otherwise removes the incremental
code entirely.