defer setExtraData in graph rebuild (fixes #117)
This primarily fixes a crash in yajilin-regions, but should generally
make graph rebuilding safer. The crash:

1. answer clear causes a call to ui.puzzle.board.roommgr.rebuild
2. this resets cell.room, and calls setComponentInfo on components
   one by one as they're found
3. calling setComponentInfo on a component eventually calls setExtraData,
   which calls component.checkAutoCmp, which calls clist.redraw(),
   which redraws the whole bounding box of clist. But the bounding box may
   contain cells from components that haven't been rebuilt yet, so haven't
   had setComponentInfo called yet which would set the room ref.

This changes the rebuild code to set component refs directly, but collect
the components and only call setExtraData on them in the very end.

(Another good option would be to not randomly call out to the painter to
redraw parts of the puzzle, but instead to queue up redraw requests and
run them in batch when it's safe (e.g. after the rebuild). But this works
for now.)