shwolf: fix encoding of 1xn and nx1 grids
The encoding of the dots (encodeCrossMarks) was emitting an empty
string, while the decoding was consuming one character.

The change encodes the crosses of such a grid as "0", which is
consistent in that absence of crosses is generally encoded as the
number of potential crosses, i.e., all crosses are skipped.

The only other puzzle type that uses encodeCrossMarks is bdblock,
which uses it with `cross: 2` hence is not affected.