Bug and enhancement for Bosanowa
Made editing more nice, by disallowing the cursor to go on borders in edit mode. Fixed a bug where numbers above 255 were not encoded. Fixed a bug where going in answer mode on a cell that is not a circle, in number mode, then clicking the cell would place a number.